// src/app/shared/models/employee.ts
export interface Iemployee {
    fname: string
    lname: string
    contact: string
    email: string
    currentAddress: string
    permentAddress: string
    city: string
    state: string
    pincode: string
    bankName: string
    ifscCode: string
    salary: string
    id?: string
    srNum: number
}

export interface UserData {
    id: string;
    name: string;
    progress: string;
    fruit: string;
}